pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama"
version = "1.0.0"
description = "Elite local AI development platform for LLM inference"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "kushin77"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "redis>=5.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "qdrant-client>=1.16.0",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "firebase-admin>=6.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "deprecated>=1.2.13",
    "prometheus-client>=0.17.0",
    "transformers>=4.53.0",
    "sentence-transformers>=2.2.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.25.0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "opentelemetry-instrumentation-httpx>=0.40b0",
    "opentelemetry-instrumentation-redis>=0.40b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.40b0",
    # PMO Agent Dependencies
    "pmo-agent-remediation>=1.0.0",
    "pmo-agent-drift-predictor>=1.0.0",
    "pmo-agent-scheduler>=1.0.0",
    "pmo-agent-audit>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "pip-audit>=2.0.0",
]

[tool.setuptools]
packages = ["ollama"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ollama --cov-report=term-missing"
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]

[tool.coverage.run]
branch = true
source = ["ollama"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
]

[tool.coverage.report]
fail_under = 90
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.pylint]
max-line-length = 100
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]
